fix: logql labels selection cleanup
